Finished build script

+++ build.py
@@ -3,27 +3,58 @@ import os
 import shutil
 import subprocess
 
+os_name = "DaedalOS"
 
 arch = "x86_64"
-foo = 1
+
+src_dir = "./src/arch/{}".format(arch)
+build_dir = "./build/arch/{}".format(arch)
+iso_dir = "./build/isofiles"
+
+flags = {"kernel": "-std=gnu99 -ffreestanding -O2 -Wall -Wextra"}
 
 
 def build():
     clean()
-    os.makedirs("./build/arch/{}".format(arch))
-    for src in os.listdir("./src/arch/{}".format(arch)):
-        if (src.split(".")[-1] == "asm"):
-            subprocess.run(["nasm", "-felf64", "./src/arch/{}/{}".format(arch, src), "-o", "./build/arch/{}/{}.o".format(arch,src)])
-            print(src)
-        # print(srcfile)
+    fnames = []
+    os.makedirs(build_dir)
+    # Build
+    for src in os.listdir(src_dir):
+        src = src.split(".")
+        name = ".".join(src[:-1])
+        if (src[-1] == "asm"):
+            subprocess.run("nasm -felf64 {0}/{2}.asm"
+                           " -o {1}/{2}.o".format(src_dir, build_dir, name),
+                           shell=True, check=True)
+            print("Assembled " + name)
+        elif (src[-1] == "c"):
+            subprocess.run("gcc -c {0}/{2}.c -o {1}/{2}.o {3}"
+                           .format(src_dir, build_dir, name, flags[name]),
+                           shell=True, check=True)
+            print("Compiled " + name)
+        else:
+            continue
+        fnames.append(build_dir + "/" + name + ".o")
+    # Link
+    subprocess.run("ld -n -o ./build/kernel.bin -T {0}/linker.ld {2}"
+                   .format(src_dir, build_dir, " ".join(fnames)),
+                   shell=True, check=True)
 
 
 def run():
-    os.makedirs("./build/isofiles/boot/grub")
+    os.makedirs(iso_dir + "/boot/grub")
+    shutil.copy(src_dir + "/grub.cfg", iso_dir + "/boot/grub/grub.cfg")
+    shutil.copy("./build/kernel.bin", iso_dir + "/boot/kernel.bin")
+    subprocess.run("grub-mkrescue -o ./build/{0}.iso {1}"
+                   .format(os_name, iso_dir), shell=True, check=True)
+    shutil.rmtree(iso_dir)
+    subprocess.run("qemu-system-x86_64 -cdrom ./build/{}.iso".format(os_name),
+                   shell=True, check=True)
 
 
 def clean():
-    shutil.rmtree("./build/")
+    if os.path.exists("./build/"):
+        shutil.rmtree("./build/")
 
 
 parser = argparse.ArgumentParser()
